requirements.txt
langchain-community==0.0.34
streamlit==1.33.0
langchain==0.1.16
chromadb==0.5.0
pypdf==4.2.0
docx2txt==0.8
psutil==5.9.8
GPUtil==1.4.0
sentence-transformers==2.6.1
InstructorEmbedding==1.0.1
scikit-learn==1.4.2
pdfplumber==0.11.0

factsumm~=0.1.5
numpy~=1.26.4